Refactor transaction handling with `TransactionMixin`.
Abstracted transaction management into a reusable `TransactionMixin` to enhance modularity and eliminate duplicate code in DAOs and DAOFactory. Updated docs to reflect new transaction features and added support for multi-DAO transactions. Adjusted tests to account for improved error handling during commits and rollbacks.
